feat(optimizer): correct errors in get_parameters_and_gradients method and finalize project.
This commit changes the return iter in get_parameters_and_gradients methods for
yield from. The optimization step and logging is finalized, and a trivial dataset
option is included in the tokenizer for testing.

Final tests with whole dataset are run to fully verify the whole Transformer
implementation. Everything passes.